gen: Move the InstanceExt generator to an instance_ext.rs module
